Cleaned up some pointer aliasing problems and some signed/unsigned compare problems.
git-svn-id: file:///home/cminyard/t2/svn/openipmi/trunk@1876 a860ddb0-99b4-4eb6-8282-105f033f2ff5

+++ libedit/read.c
@@ -487,7 +487,7 @@ el_gets(EditLine *el, int *nread)
 #endif /* DEBUG_READ */
 			break;
 		}
-		if ((unsigned int)cmdnum >= el->el_map.nfunc) {	/* BUG CHECK command */
+		if (cmdnum >= el->el_map.nfunc) {	/* BUG CHECK command */
 #ifdef DEBUG_EDIT
 			(void) fprintf(el->el_errfile,
 			    "ERROR: illegal command from key 0%o\r\n", ch);
